+++ Figurenmanager2/src/main/java/at/htl/figurenmanager2/Point.java
@@ -1,9 +1,9 @@
 package at.htl.figurenmanager2;
 
 public class Point {
-    public int x, y;
+    public float x, y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
